Segregatte AP methods
