+++ tests/suite/test_app_protect_grpc.py
@@ -1,7 +1,7 @@
 import grpc
 import pytest
 from settings import TEST_DATA, DEPLOYMENTS
-from suite.custom_resources_utils import (
+from suite.ap_resources_utils import (
     create_ap_logconf_from_yaml,
     create_ap_policy_from_yaml,
     delete_ap_policy,
