+++ tests/suite/test_app_protect_integration.py
@@ -4,12 +4,6 @@ import yaml
 import json
 
 from settings import TEST_DATA, DEPLOYMENTS
-from suite.custom_resources_utils import (
-    create_ap_logconf_from_yaml,
-    create_ap_policy_from_yaml,
-    delete_ap_policy,
-    delete_ap_logconf,
-)
 from suite.resources_utils import (
     wait_before_test,
     create_example_app,
@@ -32,7 +26,11 @@ from suite.resources_utils import (
     scale_deployment,
     get_pods_amount,
 )
-from suite.custom_resources_utils import (
+from suite.ap_resources_utils import (
+    create_ap_logconf_from_yaml,
+    create_ap_policy_from_yaml,
+    delete_ap_policy,
+    delete_ap_logconf,
     read_ap_custom_resource,
     create_ap_usersig_from_yaml,
     delete_and_create_ap_policy_from_yaml,
