+++ tests/suite/test_batch_startup_times.py
@@ -4,11 +4,14 @@ import requests
 import pytest
 import yaml
 
-from suite.custom_resources_utils import (
+from suite.ap_resources_utils import (
+    create_ap_usersig_from_yaml,
+    delete_ap_usersig,
     create_ap_logconf_from_yaml,
     create_ap_policy_from_yaml,
     delete_ap_policy,
     delete_ap_logconf,
+    create_ap_waf_policy_from_yaml,
 )
 from suite.resources_utils import (
     ensure_connection_to_public_endpoint,
@@ -33,11 +36,8 @@ from suite.resources_utils import (
 from suite.custom_resources_utils import (
     create_virtual_server_from_yaml,
     delete_virtual_server,
-    create_ap_usersig_from_yaml,
-    delete_ap_usersig,
     patch_virtual_server_from_yaml,
     create_policy_from_yaml,
-    create_ap_waf_policy_from_yaml,
     delete_policy,
     create_virtual_server,
     create_v_s_route,
